Data Intelligence Application/requirements.txt
# Core Dependencies
streamlit>=1.29.0
pandas>=2.0.0
numpy>=1.24.0

# Data Analysis & AI
pandasai>=1.5.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Database Connectors
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.1.0  # MySQL
pyodbc>=4.0.0  # SQL Server
sqlalchemy>=2.0.0  # SQL toolkit

# File Format Support
openpyxl>=3.1.0  # Excel support
xlrd>=2.0.0  # Excel reading
pyarrow>=13.0.0  # Parquet support

# API & Web
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
